flake/the_wiz/requirements.txt
-i https://pypi.org/simple
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2.1
decorator==4.4.2
flask==1.1.2
google-auth==1.21.0
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
jinja2==2.11.2
k8-kat==0.0.156
kubernetes==11.0.0
markupsafe==1.1.1
nectwiz==0.0.11
oauthlib==3.1.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.24.0
rsa==4.6
six==1.12.0
typing-extensions==3.7.4.3
urllib3==1.25.10
validators==0.18.0
websocket-client==0.57.0
werkzeug==1.0.1
